Use new preview API to consolidate Open AI logic.
